Add files via upload
 Online Railway Ticket Booking System
Railway Ticket Booking System is easy to ticket booking and information about train timing using HTML, CSS 
and PHP Online Railway Ticket Reservation System is a Web based application that works within a centralized 
network. This project presents a review on the software program "Online Railway Ticket Reservation System" as 
should be used in a train transportation system, a facility which is used to reserve seats, cancellation of reservation 
and different types of route enquiries used on securing quick reservations.